+++ Iocp20Coroutine/WorldSvr/WorldSessionFromGame.h
@@ -27,8 +27,8 @@ public:
 	WorldSvrAcceptGame* m_pServer = nullptr;
 	template<class T> std::deque<T>& GetQueue();
 	bool Process();
-	uint32_t m_snRecv = 0;
-	uint32_t m_snSend = 0;
+	uint8_t m_snRecv = 0;
+	uint8_t m_snSend = 0;
 	CompeletionKeySession& m_refSession;
 private:
 	void OnRecvPack(const void* buf, int len);
